Change how --dynamic works
This is a backwards-incompatible change.  I'm sorry.  Hey, it wasn't 1.0.0 yet.

Previously --dynamic made peat look for the listing command on standard in.
This was consistent in the sense that it was always looking for "the input
files" on standard in, but in practice it was a giant pain in the ass because it
required you to quote shell commands in two separate ways (one for the listing
command, the other for the action command).

Now you can quote both commands in the same way and avoid going insane.